+++ app/src/main/java/com/example/alldroid/QuizPackage/QuizEndFragment.java
@@ -1,4 +1,4 @@
-package com.example.alldroid;
+package com.example.alldroid.QuizPackage;
 
 import android.os.Bundle;
 
@@ -8,13 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.alldroid.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link QuizStart#newInstance} factory method to
+ * Use the {@link QuizEndFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class QuizStart extends Fragment {
+public class QuizEndFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -24,7 +26,7 @@ public class QuizStart extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public QuizStart() {
+    public QuizEndFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class QuizStart extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment QuizStart.
+     * @return A new instance of fragment QuizEndFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static QuizStart newInstance(String param1, String param2) {
-        QuizStart fragment = new QuizStart();
+    public static QuizEndFragment newInstance(String param1, String param2) {
+        QuizEndFragment fragment = new QuizEndFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class QuizStart extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_quiz_start, container, false);
+        return inflater.inflate(R.layout.fragment_quiz_end, container, false);
     }
 }
