+++ app/src/main/res/layout/fragment_quiz_end.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QuizEndFragment">
+    tools:context=".QuizPackage.QuizEndFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
