+++ app/src/main/res/layout/fragment_quiz_start.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QuizStart">
+    tools:context=".QuizPackage.QuizStart">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
